pyproject.toml
[project]
name = "archon"
version = "0.0.7"
description = "Archon Agent Framework"
readme = "README.md"
requires-python = "3.11.6"

[tool.poetry]
name = "archon"
version = "0.0.7"
description = "Archon Agent Framework"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "archon"}]

[tool.poetry.dependencies]
aiofiles="24.1.0"
aiohappyeyeballs="2.4.4"
aiohttp="3.11.11"
aiosignal="1.3.2"
aiosqlite="0.20.0"
altair="5.5.0"
annotated-types="0.7.0"
anthropic="0.49.0"
anyio="4.8.0"
argcomplete="3.6.2"
asgiref="3.8.1"
attrs="24.3.0"
beautifulsoup4="4.12.3"
blinker="1.9.0"
boto3="1.37.33"
botocore="1.37.33"
cachetools="5.5.0"
certifi="2024.12.14"
cffi="1.17.1"
charset-normalizer="3.4.1"
click="8.1.8"
cohere="5.13.12"
colorama="0.4.6"
Crawl4AI="0.4.247"
cryptography="43.0.3"
Deprecated="1.2.15"
deprecation="2.1.0"
distro="1.9.0"
dnspython="2.7.0"
docstring_parser="0.16"
email_validator="2.2.0"
eval_type_backport="0.2.2"
executing="2.1.0"
fake-http-header="0.3.5"
fastapi="0.115.8"
fastapi-cli="0.0.7"
fastavro="1.10.0"
filelock="3.16.1"
frozenlist="1.5.0"
fsspec="2024.12.0"
gitdb="4.0.12"
GitPython="3.1.44"
google-auth="2.37.0"
googleapis-common-protos="1.66.0"
gotrue="2.11.1"
greenlet="3.1.1"
griffe="1.5.4"
groq="0.15.0"
h11="0.14.0"
h2="4.1.0"
hpack="4.0.0"
html2text="2024.2.26"
httpcore="1.0.7"
httptools="0.6.4"
httpx="0.27.2"
httpx-sse="0.4.0"
huggingface-hub="0.27.1"
hyperframe="6.0.1"
idna="3.10"
importlib_metadata="8.5.0"
iniconfig="2.0.0"
instructor="1.7.9"
itsdangerous="2.2.0"
Jinja2="3.1.5"
jiter="0.8.2"
jmespath="1.0.1"
joblib="1.4.2"
jsonpatch="1.33"
jsonpath-python="1.0.6"
jsonpointer="3.0.0"
jsonschema="4.23.0"
jsonschema-specifications="2024.10.1"
jsonschema_rs="0.25.1"
langchain-core="0.3.33"
langgraph="0.3.29"
langgraph-checkpoint="2.0.24"
langgraph-cli="0.2.3"
langgraph-sdk="0.1.61"
langsmith="0.3.30"
litellm="1.57.8"
logfire={extras = ["aiohttp"], version = "^3.14.0"}
logfire-api="3.1.0"
lxml="5.3.0"
markdown-it-py="3.0.0"
MarkupSafe="3.0.2"
mcp="1.6.0"
mdurl="0.1.2"
mistralai="1.2.6"
mockito="1.5.3"
msgpack="1.1.0"
multidict="6.1.0"
mypy-extensions="1.0.0"
narwhals="1.21.1"
nltk="3.9.1"
numpy="2.2.1"
openai="1.73.0"
opentelemetry-api="1.29.0"
opentelemetry-exporter-otlp-proto-common="1.29.0"
opentelemetry-exporter-otlp-proto-http="1.29.0"
opentelemetry-instrumentation="0.50b0"
opentelemetry-instrumentation-asgi="0.50b0"
opentelemetry-instrumentation-dbapi="0.50b0"
opentelemetry-instrumentation-fastapi="0.50b0"
opentelemetry-instrumentation-sqlite3="0.50b0"
opentelemetry-proto="1.29.0"
opentelemetry-sdk="1.29.0"
opentelemetry-semantic-conventions="0.50b0"
opentelemetry-util-http="0.50b0"
orjson="3.10.15"
packaging="24.2"
pandas="2.2.3"
pillow="10.4.0"
pinecone="6.0.2"
pinecone-plugin-interface="0.0.7"
playwright="1.49.1"
pluggy="1.5.0"
postgrest="0.19.1"
prompt_toolkit="3.0.50"
propcache="0.2.1"
protobuf="5.29.3"
psutil="6.1.1"
pyarrow="18.1.0"
pyasn1="0.6.1"
pyasn1_modules="0.4.1"
pycparser="2.22"
pydantic="2.10.5"
pydantic-ai="0.0.54"
pydantic-ai-slim="0.0.54"
pydantic-evals="0.0.54"
pydantic-extra-types="2.10.2"
pydantic-graph="0.0.54"
pydantic-settings="2.7.1"
pydantic_core="2.27.2"
pydeck="0.9.1"
pyee="12.0.0"
Pygments="2.19.1"
PyJWT="2.10.1"
pyOpenSSL="24.3.0"
pytest="8.3.4"
pytest-mockito="0.0.4"
python-dateutil="2.9.0.post0"
python-dotenv="1.0.1"
python-multipart="0.0.20"
pytz="2024.2"
PyYAML="6.0.2"
rank-bm25="0.2.2"
realtime="2.1.0"
referencing="0.35.1"
regex="2024.11.6"
requests="2.32.3"
requests-toolbelt="1.0.0"
rich="13.9.4"
rich-toolkit="0.13.2"
rpds-py="0.22.3"
rsa="4.9"
s3transfer="0.11.4"
shellingham="1.5.4"
six="1.17.0"
smmap="5.0.2"
sniffio="1.3.1"
snowballstemmer="2.2.0"
soupsieve="2.6"
sse-starlette="2.1.3"
starlette="0.45.3"
storage3="0.11.0"
streamlit="1.41.1"
StrEnum="0.4.15"
structlog="24.4.0"
supabase="2.11.0"
supafunc="0.9.0"
tenacity="9.0.0"
tf-playwright-stealth="1.1.0"
tiktoken="0.8.0"
tokenizers="0.21.0"
toml="0.10.2"
tornado="6.4.2"
tqdm="4.67.1"
typer="0.15.1"
types-requests="2.32.0.20241016"
typing-inspect="0.9.0"
typing-inspection="0.4.0"
typing_extensions="4.12.2"
tzdata="2024.2"
ujson="5.10.0"
urllib3="2.3.0"
uvicorn="0.34.0"
uvloop="0.21.0"
watchdog="6.0.0"
watchfiles="1.0.4"
wcwidth="0.2.13"
websockets="13.1"
wrapt="1.17.1"
xxhash="3.5.0"
yarl="1.18.3"
zipp="3.21.0"
zstandard="0.23.0"
langgraph-api = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 